add: Move properties handler to a trait.
add: :new: trait called NelsonMartell\PropertiesHandler
